add date-now option and fix renamePhoto args
The --date-now option allows for files without a date taken value to
still be renamed using the current date. With this change, a file has
EXIF data if the colorspace can be read using exiftool.

Made the array or organiz extensions last so that the descArg string
could be passed as a value